# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для биллинга",
        "description": "Этот API предоставляет систему биллинга, в которой пользователи могут зарегистрироваться, проверять свой баланс, выполнять транзакции (пополнение/снятие) и просматривать историю транзакций. Все операции проходят проверку, исключая возможность отрицательного баланса.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер для разработки API биллинга."
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Этот метод позволяет зарегистрировать нового пользователя в системе.",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя пользователя (от 3 до 20 символов)",
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "description": "Уникальный email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль (минимум 8 символов)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messaged": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Registered"
                                        },
                                        "token": {
                                            "description": "Токен аутентификации",
                                            "type": "string",
                                            "example": "3|QRkUnUf2syjUmOyvSiO8p6hjaZaeOnHJPuwgkunL7f07d3b2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string",
                                            "example": "Поле email является обязательным."
                                        },
                                        "errors": {
                                            "description": "Детализированные ошибки валидации",
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "Поле email является обязательным."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Вход пользователя в систему",
                "description": "Этот метод используется для входа пользователя в систему с использованием email и пароля.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль пользователя (минимум 8 символов)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messaged": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Logged In"
                                        },
                                        "token": {
                                            "description": "Токен аутентификации",
                                            "type": "string",
                                            "example": "3|QRkUnUf2syjUmOyvSiO8p6hjaZaeOnHJPuwgkunL7f07d3b2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messaged": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "The email or password is incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string",
                                            "example": "Поле email является обязательным."
                                        },
                                        "errors": {
                                            "description": "Детализированные ошибки валидации",
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "Поле email является обязательным."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Используйте Bearer Token для аутентификации. Все защищенные маршруты требуют действительного токена после входа пользователя в систему.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}